refactor: add utils and moved some functions there in sorting-page
